requirements.txt
playwright
playwright-stealth
langgraph
langchain-anthropic
anthropic
typer
rich
python-dotenv
